Prime Factors Math Challenge
Leonardo loves primes and created q queries where each query takes the form of an integer, n. For each n, he wants you to count the maximum number of unique prime factors of any number in the inclusive range [1,n] and then print this value on a new line.

Note: Recall that a prime number is only divisible by 1 and itself, and 1 is not a prime number.

Input Format

The first line contains an integer, q, denoting the number of queries. 
Each line i of the q subsequent lines contains a single integer, n .

Written in Python 3.